test4

+++ main.py
@@ -79,7 +79,7 @@ def main():
     detector.load_llm(
         llm_name=LLM_NAME,
         use_local=False,
-        dtype=torch.bfloat16,  # Usa float16 o bfloat16 (rtx30) a seconda della tua GPU (rtx20)
+        dtype=torch.float16,  # Usa float16 o bfloat16 (rtx30) a seconda della tua GPU (rtx20)
         use_device_map=True
     )
